[chore] 부하 테스트 패키지 이동 (#64)
