+++ src/test/java/hyundai/softeer/orange/load/RedisLuaFcfsServiceLoadTest.java
@@ -1,16 +1,13 @@
-package hyundai.softeer.orange.event.fcfs.load;
-
-import hyundai.softeer.orange.event.fcfs.service.RedisSetFcfsService;
+package hyundai.softeer.orange.load;
+import hyundai.softeer.orange.event.fcfs.service.RedisLuaFcfsService;
 import hyundai.softeer.orange.event.fcfs.util.FcfsUtil;
 import lombok.extern.slf4j.Slf4j;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.core.StringRedisTemplate;
-
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
@@ -19,10 +16,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @Slf4j
 @SpringBootTest
-class RedisSetFcfsServiceLoadTest {
+class RedisLuaFcfsServiceLoadTest {
 
     @Autowired
-    private RedisSetFcfsService redisSetFcfsService;
+    RedisLuaFcfsService redisLuaFcfsService;
 
     @Autowired
     StringRedisTemplate stringRedisTemplate;
@@ -44,17 +41,8 @@ class RedisSetFcfsServiceLoadTest {
         booleanRedisTemplate.getConnectionFactory().getConnection().flushAll();
 
         // 테스트할 이벤트 정보 저장
-        booleanRedisTemplate.opsForValue().set(FcfsUtil.endFlagFormatting(eventSequence.toString()), false);
         numberRedisTemplate.opsForValue().set(FcfsUtil.keyFormatting(eventSequence.toString()), numberOfWinners);
         stringRedisTemplate.opsForValue().set(FcfsUtil.startTimeFormatting(eventSequence.toString()), "2021-08-01T00:00:00");
-    }
-
-    @AfterEach
-    void tearDown() {
-        // 초기화
-        stringRedisTemplate.getConnectionFactory().getConnection().flushAll();
-        numberRedisTemplate.getConnectionFactory().getConnection().flushAll();
-        booleanRedisTemplate.getConnectionFactory().getConnection().flushAll();
     }
 
     @Test
@@ -70,7 +58,7 @@ class RedisSetFcfsServiceLoadTest {
             final int index = i;
             executorService.execute(() -> {
                 try {
-                    boolean result = redisSetFcfsService.participate(1L, "user" + index);
+                    boolean result = redisLuaFcfsService.participate(1L, "user" + index);
                 } catch (Exception e) {
                     e.printStackTrace();
                 } finally {
